# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/auth/signup": {
      "post": {
        "operationId": "AuthController_signup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OTP sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "OTP sent successfully. Please verify to complete registration."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid phone number format"
          },
          "409": {
            "description": "Account already exists"
          }
        },
        "summary": "Request signup OTP for new user",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/signup/verify": {
      "post": {
        "operationId": "AuthController_signupVerify",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupVerifyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Account created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number"
                        },
                        "phone": {
                          "type": "string"
                        },
                        "fullName": {
                          "type": "string"
                        },
                        "isVerified": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired OTP"
          },
          "409": {
            "description": "Account already exists"
          }
        },
        "summary": "Verify OTP and create new account",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/signin": {
      "post": {
        "operationId": "AuthController_signin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SigninDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "OTP sent successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid phone number format"
          },
          "404": {
            "description": "No account found with this phone number"
          }
        },
        "summary": "Request signin OTP for existing user",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/signin/verify": {
      "post": {
        "operationId": "AuthController_signinVerify",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SigninVerifyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signed in successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number"
                        },
                        "phone": {
                          "type": "string"
                        },
                        "fullName": {
                          "type": "string"
                        },
                        "isVerified": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired OTP"
          },
          "404": {
            "description": "User not found"
          }
        },
        "summary": "Verify OTP and sign in",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current user details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "phone": {
                      "type": "string"
                    },
                    "fullName": {
                      "type": "string",
                      "nullable": true
                    },
                    "isVerified": {
                      "type": "boolean"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current authenticated user",
        "tags": [
          "auth"
        ]
      }
    }
  },
  "info": {
    "title": "Complaints API",
    "description": "Consumer complaint resolution platform API for Macedonia/Albania",
    "version": "1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "auth",
      "description": "Authentication endpoints"
    },
    {
      "name": "users",
      "description": "User management"
    },
    {
      "name": "companies",
      "description": "Company management"
    },
    {
      "name": "complaints",
      "description": "Complaint management"
    },
    {
      "name": "categories",
      "description": "Category management"
    }
  ],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "SignupDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "pattern": "/^\\+[1-9]\\d{1,14}$/",
            "example": "+38970123456",
            "description": "Phone number in E.164 format"
          },
          "fullName": {
            "type": "string",
            "minLength": 2,
            "example": "John Doe",
            "description": "Full name of the user"
          }
        },
        "required": [
          "phone",
          "fullName"
        ]
      },
      "SignupVerifyDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "pattern": "/^\\+[1-9]\\d{1,14}$/",
            "example": "+38970123456",
            "description": "Phone number in E.164 format"
          },
          "code": {
            "type": "string",
            "pattern": "/^\\d{6}$/",
            "example": "123456",
            "description": "6-digit OTP code"
          }
        },
        "required": [
          "phone",
          "code"
        ]
      },
      "SigninDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "pattern": "/^\\+[1-9]\\d{1,14}$/",
            "example": "+38970123456",
            "description": "Phone number in E.164 format"
          }
        },
        "required": [
          "phone"
        ]
      },
      "SigninVerifyDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "pattern": "/^\\+[1-9]\\d{1,14}$/",
            "example": "+38970123456",
            "description": "Phone number in E.164 format"
          },
          "code": {
            "type": "string",
            "pattern": "/^\\d{6}$/",
            "example": "123456",
            "description": "6-digit OTP code"
          }
        },
        "required": [
          "phone",
          "code"
        ]
      }
    }
  }
}